+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+- [Fixed error when calling `fungibleFromBigInteger` with invalid input](https://github.com/multiversx/mx-sdk-dapp/pull/1221)
 - [Prevent network update if there are no changes](https://github.com/multiversx/mx-sdk-dapp/pull/1220)
 
 ## [[v2.36.2]](https://github.com/multiversx/mx-sdk-dapp/pull/1219)] - 2024-08-06